code cleanup and documentation expansion continued
